Fix poor resource management code (fixes #179)
When the response contents are copied the input stream was not being
closed. This might appear correct for a one-off tests due to new
connections being established. Unfortunately it eventually consumes
the entire pool with stuck connections, resulting in no new requests
being sent. Cancelling the request isn't enough since the connection
is still stuck in a bad state.

The simple solution is to close the InputStream, but since OkHttp has
a method to do this we'll just delete this bad code entirely.