Added support to Place IDs
Place IDs can be provided as origin, destination and waypoints. But in order to be used, they must be prefixed with `place_id:`.
There's a new helped method for prefixing Place IDs, which could be used to pass Place ID waypoints into the waypoints method that accepts Strings. This way users can mix addresses and Place IDs if they want to.